Reuse SignUpLink and LoginLink
