+++ packages/lesmis/src/components/NavBar.js
@@ -6,9 +6,7 @@
 import React from 'react';
 import styled from 'styled-components';
 import PropTypes from 'prop-types';
-import { Link as RouterLink, useHistory } from 'react-router-dom';
 import MuiContainer from '@material-ui/core/Container';
-import MuiButton from '@material-ui/core/Button';
 import { HomeButton } from '@tupaia/ui-components';
 import { FlexSpaceBetween, FlexStart } from './Layout';
 import { ProfileButton } from './ProfileButton';
@@ -16,7 +14,8 @@ import { MainMenu } from './MainMenu';
 import { SearchBar } from './SearchBar';
 import { NAVBAR_HEIGHT } from '../constants';
 import { useUser } from '../api';
-import { useUrlParams, useHomeUrl, I18n } from '../utils';
+import { useUrlParams, useHomeUrl } from '../utils';
+import { SignUpLink } from './SignUpLink';
 
 const Container = styled.nav`
   position: sticky;
@@ -39,11 +38,6 @@ const Left = styled(FlexStart)`
   height: 100%;
 `;
 
-const TextButton = styled(MuiButton)`
-  margin-right: 10px;
-  color: white;
-`;
-
 const Search = styled(SearchBar)`
   @media screen and (max-width: 900px) {
     display: none;
@@ -51,9 +45,8 @@ const Search = styled(SearchBar)`
 `;
 
 export const NavBar = ({ hideSearch }) => {
-  const history = useHistory();
   const { isLoggedIn } = useUser();
-  const { locale, view } = useUrlParams();
+  const { view } = useUrlParams();
   const { homeUrl } = useHomeUrl();
 
   return (
@@ -67,15 +60,7 @@ export const NavBar = ({ hideSearch }) => {
           {!hideSearch && <Search linkType={view} />}
           <FlexStart>
             {isLoggedIn ? null : ( // @see https://github.com/beyondessential/tupaia-backlog/issues/2290 //Todo: add Favourites Menu
-              <TextButton
-                to={{
-                  pathname: `/${locale}/register`,
-                  state: { referer: history.location },
-                }}
-                component={RouterLink}
-              >
-                <I18n t="home.signUp" />
-              </TextButton>
+              <SignUpLink />
             )}
             <ProfileButton />
           </FlexStart>
