+++ packages/lesmis/src/components/index.js
@@ -18,6 +18,7 @@ export * from './FullPageLoader';
 export * from './Layout';
 export * from './LocaleMenu';
 export * from './LocationHeader';
+export * from './LoginLink';
 export * from '../utils/I18n';
 export * from './MapOverlaysPanel';
 export * from './MiniMap';
@@ -27,6 +28,7 @@ export * from './PanelComponent';
 export * from './RejectButton';
 export * from './SearchBar';
 export * from './ScrollToTopButton';
+export * from './SignUpLink';
 export * from './Tabs';
 export * from './ThemeTest';
 export * from './VitalsLoader';
